% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_all.R
\name{read_all}
\alias{read_all}
\alias{read_all.default}
\title{read_all: Read all files in directory on a FileSystem}
\usage{
read_all(fs, dirpath, ...)

\method{read_all}{default}(fs, dirpath, ...)
}
\arguments{
\item{fs}{FileSystem object}

\item{dirpath}{a character vector that contains the path of files to read}

\item{...}{other arguments}
}
\value{
a character vector that contains the contents of the File
}
\description{
Read all Files in a directory on a FileSystem
}

